PR feedback - move node lambda from lib to ingest folder
