+++ backend/compact-connect/lambdas/nodejs/ingest-event-reporter/handler.ts
@@ -1,6 +1,6 @@
 import { DynamoDBClient } from '@aws-sdk/client-dynamodb';
 import { SESClient } from '@aws-sdk/client-ses';
-import { Lambda } from '../lib/lambda';
+import { Lambda } from './lambda';
 
 
 const lambda = new Lambda({
